requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
insightface>=0.7.3
onnxruntime-gpu>=1.20.0
opencv-python>=4.10.0
httpx>=0.28.0
python-multipart>=0.0.20
numpy>=1.24,<2.0
Pillow>=10.0
python-dotenv>=1.0.0
